requirements.txt
appnope==0.1.2
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
decorator==5.1.0
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
gunicorn==20.1.0
idna==3.2
ipython==7.27.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.24
traitlets==5.1.0
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.1
WTForms==2.3.3
